make it possible to run multiple jackals in gazebo
There is a branch for this already but it's 5 years old so better to start again.

Some changes to topics to move them to relative rather than absolute names (i.e. removing prefixed /)

Move some frame definitions out of the robot_localization file to allow them to be correctly prefixed with the robot name

Pass the robot name to some of the launch files which require it so that the tf prefixing works

Remove robotNamespace tags from jackal.gazebo so that the name of the robot model is used as the namespace rather than /