no leading slash on unprefixed frames
Workaround in control.launch because can't use an empty string for an if
attribute, so need the tf_prefix that is received to have a trailing slash for
valid frames.

Modifications to jackal.gazebo to allow it to take tf_prefix as a param so that
it can be used in if/unless statements

Both these changes should mean that all frames have no leading slash, which some
modules complain about in a load of error spam.